+++ contracts/SafeMath.sol
@@ -49,8 +49,8 @@ library SafeMath {
    *
    * - Addition cannot overflow.
    */
-  function add(uint256 a, uint256 b) internal pure returns (uint256) {
-    uint256 c = a + b;
+  function add(uint a, uint b) internal pure returns (uint) {
+    uint c = a + b;
     require(c >= a, "SafeMath: addition overflow");
 
     return c;
@@ -66,7 +66,7 @@ library SafeMath {
    *
    * - Subtraction cannot overflow.
    */
-  function sub(uint256 a, uint256 b) internal pure returns (uint256) {
+  function sub(uint a, uint b) internal pure returns (uint) {
     return sub(a, b, "SafeMath: subtraction overflow");
   }
 
@@ -81,12 +81,12 @@ library SafeMath {
    * - Subtraction cannot overflow.
    */
   function sub(
-    uint256 a,
-    uint256 b,
+    uint a,
+    uint b,
     string memory errorMessage
-  ) internal pure returns (uint256) {
+  ) internal pure returns (uint) {
     require(b <= a, errorMessage);
-    uint256 c = a - b;
+    uint c = a - b;
 
     return c;
   }
@@ -101,7 +101,7 @@ library SafeMath {
    *
    * - Multiplication cannot overflow.
    */
-  function mul(uint256 a, uint256 b) internal pure returns (uint256) {
+  function mul(uint a, uint b) internal pure returns (uint) {
     // Gas optimization: this is cheaper than requiring 'a' not being zero, but the
     // benefit is lost if 'b' is also tested.
     // See: https://github.com/OpenZeppelin/openzeppelin-contracts/pull/522
@@ -109,7 +109,7 @@ library SafeMath {
       return 0;
     }
 
-    uint256 c = a * b;
+    uint c = a * b;
     require(c / a == b, "SafeMath: multiplication overflow");
 
     return c;
@@ -127,7 +127,7 @@ library SafeMath {
    *
    * - The divisor cannot be zero.
    */
-  function div(uint256 a, uint256 b) internal pure returns (uint256) {
+  function div(uint a, uint b) internal pure returns (uint) {
     return div(a, b, "SafeMath: division by zero");
   }
 
@@ -144,12 +144,12 @@ library SafeMath {
    * - The divisor cannot be zero.
    */
   function div(
-    uint256 a,
-    uint256 b,
+    uint a,
+    uint b,
     string memory errorMessage
-  ) internal pure returns (uint256) {
+  ) internal pure returns (uint) {
     require(b > 0, errorMessage);
-    uint256 c = a / b;
+    uint c = a / b;
     // assert(a == b * c + a % b); // There is no case in which this doesn't hold
 
     return c;
@@ -167,7 +167,7 @@ library SafeMath {
    *
    * - The divisor cannot be zero.
    */
-  function mod(uint256 a, uint256 b) internal pure returns (uint256) {
+  function mod(uint a, uint b) internal pure returns (uint) {
     return mod(a, b, "SafeMath: modulo by zero");
   }
 
@@ -184,10 +184,10 @@ library SafeMath {
    * - The divisor cannot be zero.
    */
   function mod(
-    uint256 a,
-    uint256 b,
+    uint a,
+    uint b,
     string memory errorMessage
-  ) internal pure returns (uint256) {
+  ) internal pure returns (uint) {
     require(b != 0, errorMessage);
     return a % b;
   }
