+++ contracts/interfaces.sol
@@ -3,17 +3,17 @@ pragma solidity ^0.7.0;
 
 interface IMaker {
   function execute(
-    uint256 takerWants,
-    uint256 takerGives,
-    uint256 orderPenaltyPerGas,
-    uint256 orderId
+    uint takerWants,
+    uint takerGives,
+    uint orderPenaltyPerGas,
+    uint orderId
   ) external;
 }
 
 interface ITaker {
-  function take(uint256 orderId, uint256 wants) external;
+  function take(uint orderId, uint wants) external;
 
-  function mo(uint256 wants, uint256 gives) external;
+  function mo(uint wants, uint gives) external;
 }
 
 // IERC20 From OpenZeppelin code
@@ -25,12 +25,12 @@ interface IERC20 {
   /**
    * @dev Returns the amount of tokens in existence.
    */
-  function totalSupply() external view returns (uint256);
+  function totalSupply() external view returns (uint);
 
   /**
    * @dev Returns the amount of tokens owned by `account`.
    */
-  function balanceOf(address account) external view returns (uint256);
+  function balanceOf(address account) external view returns (uint);
 
   /**
    * @dev Moves `amount` tokens from the caller's account to `recipient`.
@@ -39,7 +39,7 @@ interface IERC20 {
    *
    * Emits a {Transfer} event.
    */
-  function transfer(address recipient, uint256 amount) external returns (bool);
+  function transfer(address recipient, uint amount) external returns (bool);
 
   /**
    * @dev Returns the remaining number of tokens that `spender` will be
@@ -51,7 +51,7 @@ interface IERC20 {
   function allowance(address owner, address spender)
     external
     view
-    returns (uint256);
+    returns (uint);
 
   /**
    * @dev Sets `amount` as the allowance of `spender` over the caller's tokens.
@@ -67,7 +67,7 @@ interface IERC20 {
    *
    * Emits an {Approval} event.
    */
-  function approve(address spender, uint256 amount) external returns (bool);
+  function approve(address spender, uint amount) external returns (bool);
 
   /**
    * @dev Moves `amount` tokens from `sender` to `recipient` using the
@@ -81,7 +81,7 @@ interface IERC20 {
   function transferFrom(
     address sender,
     address recipient,
-    uint256 amount
+    uint amount
   ) external returns (bool);
 
   /**
@@ -90,11 +90,11 @@ interface IERC20 {
    *
    * Note that `value` may be zero.
    */
-  event Transfer(address indexed from, address indexed to, uint256 value);
+  event Transfer(address indexed from, address indexed to, uint value);
 
   /**
    * @dev Emitted when the allowance of a `spender` for an `owner` is set by
    * a call to {approve}. `value` is the new allowance.
    */
-  event Approval(address indexed owner, address indexed spender, uint256 value);
+  event Approval(address indexed owner, address indexed spender, uint value);
 }
